Hackathon/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
distlib==0.3.6
Django==4.1.5
djangorestframework==3.14.0
filelock==3.9.0
Pillow==9.5.0
platformdirs==2.6.2
pybase64==1.2.3
pycodestyle==2.10.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.17.1
